pyproject.toml
[tool.poetry]
name = "book-management-system-demo"
version = "0.1.0"
description = ""
authors = ["RuwiRowe <ruwindarowel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
asyncpg = "^0.30.0"
pydantic-settings = "^2.9.1"
sqlmodel = "^0.0.24"
alembic = "^1.15.2"
passlib = "1.7.4"
bcrypt = "3.2.0"
pyjwt = "^2.10.1"
redis = "^5.0.3"
setuptools = "^80.4.0"
pytest = "^8.3.5"
httpx = "^0.28.1"
schemathesis = "^3.39.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
